% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{add_textcounts}
\alias{add_textcounts}
\title{Add wordcount and character count to dataframe}
\usage{
add_textcounts(x, text_col, word = TRUE, char = TRUE)
}
\arguments{
\item{highlight}{a regex string to be highlighted}

\item{highlight_style}{Interpret regex directly or, if "stem" is specified,
highlight whole words beginning with the stem specified}
}
\description{
Add wordcount and character count to dataframe
}
\examples{
testdf <- data.frame(author=c("me", "you"), text = c("democracy is so nice pero la pobreza no es bueno", "democratic movements happen en la realidad"))
add_textcounts(testdf, text=text)
}
