% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infrequent_terms.R
\name{get_infrequent_terms}
\alias{get_infrequent_terms}
\title{this is a slight modification fo the function remove_infrequent_terms()
from the preText package}
\usage{
get_infrequent_terms(x, proportion_threshold = 0.01)
}
\arguments{
\item{x}{input: takes corpus, vector of characters, or dfm objects}

\item{proportion_threshold}{which proportion of documents should a word appear
in at minimum to prevent being removed (listed here)}
}
\description{
this is a slight modification fo the function remove_infrequent_terms()
from the preText package
}
\examples{
testdf <- data.frame(author=c("me", "you"), text = c("democracy is so nice", "democratic movements happen en la realidad"))
get_infrequent_terms(testdf$text, proportion_threshold = 0.01)
}
