% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightlining.R
\name{meanRootPairs}
\alias{meanRootPairs}
\title{Calculate the mean Root Pairs number for each respondent
over a set of supplied variables}
\usage{
meanRootPairs(x, inputVars, newVarName)
}
\arguments{
\item{inputVars}{vector of variable names}

\item{newVarName}{the name the new variable should take}
}
\description{
Calculate the mean Root Pairs number for each respondent
over a set of supplied variables
}
\examples{
to come
}
