% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiquestion_plot.R
\name{multiquestion_plot}
\alias{multiquestion_plot}
\title{Build a multi-graph plot that combines plots for each of the questions
you want to compare}
\usage{
multiquestion_plot(
  x,
  vars = NULL,
  extract_labels = TRUE,
  labels = NULL,
  fill_var = NULL,
  title_text = NULL
)
}
\arguments{
\item{vars}{a vector of column names}

\item{extract_labels}{if the columns have labels, you can extract them and
make them the title of each plot}

\item{labels}{manually specify labels}

\item{fill_var}{Variable to be used as fill aesthetic in ggplot}

\item{title_text}{overall graph title
multiquestion_plot()}
}
\description{
Build a multi-graph plot that combines plots for each of the questions
you want to compare
}
