% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_messages.R
\name{print_data}
\alias{print_data}
\title{Print a big, text-heavy dataframe to an html with optional organization
and higlighting}
\usage{
print_data(
  x,
  att1 = NULL,
  att2 = NULL,
  text = NULL,
  keywords = NULL,
  highlight = NULL,
  highlight_style = "stem"
)
}
\arguments{
\item{att1}{first attribute to be displayed in the left sidebar}

\item{att2}{second attribute to be displayed in the left sidebar}

\item{text}{which column of the dataframe contains the text to print.
Do not put this in quotes!}

\item{keywords}{string or vector of strings that are used to split the
file into sections}

\item{highlight}{a regex string to be highlighted}

\item{highlight_style}{Interpret regex directly or, if "stem" is specified,
highlight whole words beginning with the stem specified
print_data()}
}
\description{
Print a big, text-heavy dataframe to an html with optional organization
and higlighting
}
