% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{reverse}
\alias{reverse}
\title{Reverse codings of items (such as in a survey)}
\usage{
reverse(x)
}
\description{
Reverse codings of items (such as in a survey)
}
\examples{
ideology <-  c(1, 4, 4, 5, 2, 2, 2, 4, 4, 6, 2, 4, 5, 7, 4, 1, 2, 3, 2, 3, 8)
reverse(ideology)
library(dplyr)
df <- data.frame(id = seq(1:length(ideology)),ideology )
df \%>\% mutate(ideology_r = reverse(ideology))
}
