% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{add_textcounts}
\alias{add_textcounts}
\title{Add wordcount and character count to dataframe}
\usage{
add_textcounts(x, text_col, word = TRUE, char = TRUE)
}
\arguments{
\item{text_col}{which column contains text to be counted}

\item{word}{add wordcount}

\item{char}{add character count}
}
\description{
Add wordcount and character count to dataframe
}
\examples{
testdf <- data.frame(author=c("me", "you"), text = c("democracy is so nice", "democratic movements happen en la realidad"))
add_textcounts(testdf, text=text)
}
